+++ src/QueryType/Select/Result/Result.php
@@ -6,11 +6,12 @@ use Solarium\Component\ComponentAwareQueryInterface;
 use Solarium\Component\Result\Debug\Result as DebugResult;
 use Solarium\Component\Result\FacetSet as FacetSetResult;
 use Solarium\Component\Result\Grouping\Result as GroupingResult;
-use Solarium\Component\Result\Highlighting\Highlighting as HighlightingResult;
-use Solarium\Component\Result\MoreLikeThis\Result as MoreLikeThisResult;
+use Solarium\Component\Result\Highlighting\Highlighting;
+use Solarium\Component\Result\MoreLikeThis\MoreLikeThis;
 use Solarium\Component\Result\Spellcheck\Result as SpellcheckResult;
 use Solarium\Component\Result\Suggester\Result as SuggesterResult;
-use Solarium\Component\Result\Stats\Result as StatsResult;
+use Solarium\Component\Result\Stats\Stats;
+use Solarium\Core\Query\DocumentInterface;
 use Solarium\Core\Query\Result\QueryType as BaseResult;
 
 /**
@@ -232,9 +233,9 @@ class Result extends BaseResult implements \IteratorAggregate, \Countable
      *
      * This is a convenience method that maps presets to getComponent
      *
-     * @return MoreLikeThisResult|null
+     * @return MoreLikeThis|null
      */
-    public function getMoreLikeThis(): ?MoreLikeThisResult
+    public function getMoreLikeThis(): ?MoreLikeThis
     {
         return $this->getComponent(ComponentAwareQueryInterface::COMPONENT_MORELIKETHIS);
     }
@@ -244,9 +245,9 @@ class Result extends BaseResult implements \IteratorAggregate, \Countable
      *
      * This is a convenience method that maps presets to getComponent
      *
-     * @return HighlightingResult|null
+     * @return Highlighting|null
      */
-    public function getHighlighting(): ?HighlightingResult
+    public function getHighlighting(): ?Highlighting
     {
         return $this->getComponent(ComponentAwareQueryInterface::COMPONENT_HIGHLIGHTING);
     }
@@ -304,9 +305,9 @@ class Result extends BaseResult implements \IteratorAggregate, \Countable
      *
      * This is a convenience method that maps presets to getComponent
      *
-     * @return StatsResult|null
+     * @return Stats|null
      */
-    public function getStats(): ?StatsResult
+    public function getStats(): ?Stats
     {
         return $this->getComponent(ComponentAwareQueryInterface::COMPONENT_STATS);
     }